+++ libonyxia/src/error/volume_error.rs
@@ -3,44 +3,44 @@ use failure::Fail;
 #[derive(Debug, Fail)]
 pub enum VolumeError {
     #[fail(display = "Create. id: {}, cause: {}", id, cause)]
-    Create { id: usize, cause: String },
+    Create { id: u32, cause: String },
 
     #[fail(display = "ReadOnly. id: {}", _0)]
-    ReadOnly(usize),
+    ReadOnly(u32),
 
     #[fail(
         display = "Overflow. id: {}, max: {}, current: {}, todo: {}",
         id, max, current, todo
     )]
     Overflow {
-        id: usize,
+        id: u32,
         max: u64,
         current: u64,
         todo: u64,
     },
 
     #[fail(display = "Data corruption. id: {}, cause: {}", id, cause)]
-    DataCorruption { id: usize, cause: String },
+    DataCorruption { id: u32, cause: String },
 
     #[fail(
         display = "WriteLengthMismatch. id: {}, path: {}, input length: {}, receive length: {}",
         id, path, input_length, receive_length
     )]
     WriteLengthMismatch {
-        id: usize,
+        id: u32,
         path: String,
         input_length: usize,
         receive_length: usize,
     },
     #[fail(display = "Not Found. {}", _0)]
-    NotFound(usize),
+    NotFound(u32),
 
     #[fail(display = "Write Needle. path: {}, cause: {}", path, cause)]
     WriteNeedle { path: String, cause: String },
 }
 
 impl VolumeError {
-    pub fn create<C>(id: usize, cause: C) -> VolumeError
+    pub fn create<C>(id: u32, cause: C) -> VolumeError
     where
         C: Into<String>,
     {
@@ -48,11 +48,11 @@ impl VolumeError {
         VolumeError::Create { id, cause }
     }
 
-    pub fn readonly(index: usize) -> VolumeError {
+    pub fn readonly(index: u32) -> VolumeError {
         VolumeError::ReadOnly(index)
     }
 
-    pub fn overflow(id: usize, max: u64, current: u64, todo: u64) -> VolumeError {
+    pub fn overflow(id: u32, max: u64, current: u64, todo: u64) -> VolumeError {
         VolumeError::Overflow {
             id,
             max,
@@ -61,7 +61,7 @@ impl VolumeError {
         }
     }
 
-    pub fn data_corruption<C>(id: usize, cause: C) -> VolumeError
+    pub fn data_corruption<C>(id: u32, cause: C) -> VolumeError
     where
         C: Into<String>,
     {
@@ -70,7 +70,7 @@ impl VolumeError {
     }
 
     pub fn write_length_mismatch<P>(
-        id: usize,
+        id: u32,
         path: P,
         input_length: usize,
         receive_length: usize,
@@ -87,7 +87,7 @@ impl VolumeError {
         }
     }
 
-    pub fn not_found(id: usize) -> VolumeError {
+    pub fn not_found(id: u32) -> VolumeError {
         VolumeError::NotFound(id)
     }
